pyproject.toml
[project]
name = "img-move-label"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.136.0",
    "jinja2>=3.1.6",
    "pillow>=12.2.0",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
    "watchfiles>=1.1.1",
]
